requirements.txt
pandas
numpy
google-cloud-bigquery
scikit-learn
joblib
google-auth-oauthlib
google-cloud-storage
httpx
google-adk
google-cloud-texttospeech
